refactor: break out types
